Added PressureMat_v11_23x18 arduino  code
Arduino code to drive a 23 x 18 pressure sensor matrix. The application of pressure on the mat is visualized in real time in the serial monitor. Original code was obtained from a Sensitronics tutorial but most of it was modified for this specific application. Previous versions of this code were written but this is the working code for the 23 x 18 pressure sensor mat prototype that was developed.